made the requested changes
