+++ src/pages/_app.js
@@ -16,12 +16,6 @@ const MyApp = (props) => {
   const router = useRouter();
 
   const [loading, setLoading] = useState(false);
-  const [isOptionKey, setIsOptionKey] = useState(false);
-
-  const initialvalue = {
-    isOptionKey,
-    setIsOptionKey,
-  };
 
   useEffect(() => {
     const handleStart = (url) => url !== router.asPath && setLoading(true);
@@ -50,20 +44,8 @@ const MyApp = (props) => {
       <UserContextProvider>
         <MembersProvider>
           <SearchMemberProvider>
-            <KeyboardProvider value={initialvalue}>
-              <div
-                className={classNames.root}
-                role="button"
-                tabIndex={0}
-                onKeyDown={(e) => {
-                  if (e.altKey) {
-                    setIsOptionKey(true);
-                  }
-                }}
-                onKeyUp={() => {
-                  setIsOptionKey(false);
-                }}
-              >
+            <KeyboardProvider>
+              <div className={classNames.root}>
                 <div className={classNames.main}>
                   {loading && <Spinner />}
                   <Component {...pageProps} />
